feat: View refactor
Aqui a refatoração foi mais no sentido de quebrar a antiga Classe CastrarCliente em subclasses, deixando o código menor. Não houveram mudanças funcionais.
