+++ src/views/interfaces/IView.java
@@ -1,0 +1,12 @@
+package views.interfaces;
+
+public interface IView {
+    String selecionarOpcao(String menu);
+    void exibeDialogo(String mensagem);
+
+    String solicitaEntradaDeDado(String mensagem);
+
+    String solicitaEntradaDeDado(String mensagem, String valorInicial);
+
+    int confirmaAcao(String mensagem);
+}
