+++ src/views/mesa/CadastroMesaView.java
@@ -1,4 +1,4 @@
-package views;
+package views.mesa;
 
 import java.util.Scanner;
